feat - configuração de autorização de acesso e alteração de epecialidades

+++ src/main/java/com/mballem/curso/security/web/controller/MedicoController.java
@@ -58,8 +58,13 @@ public class MedicoController {
 	
 	@GetMapping({"/id/{idMed}/excluir/especializacao/{idEsp}"})
 	public String excluirEspecialidadePorMedico(@PathVariable("idMed") Long idMed, @PathVariable("idEsp") Long idEsp, RedirectAttributes attr) {
-		service.excluirEspecialidadePorMedico(idMed, idEsp);
-		attr.addFlashAttribute("sucesso", "Especialidade removida com sucesso.");
+		if(service.existeEspecialidadePorMedico(idMed,idEsp)) {
+			attr.addFlashAttribute("falha", "Tem consultas agendadas.");			
+		}else {
+			service.excluirEspecialidadePorMedico(idMed, idEsp);
+			attr.addFlashAttribute("sucesso", "Especialidade removida com sucesso.");			
+		}
+		
 		return "redirect:/medicos/dados";
 	}
